fix: calloc call by swapping arg order
The prototype of calloc is
  void *calloc(size_t nmemb, size_t size);
denoting "an array of nmemb objects, each of whose size is size." (C23)

merely changing the order of arguments to put the size at the
end (ie sizeof(char)) is what gcc seems to be looking for.
It shouldn't affect anything other than this gcc warning,
but it's good practice to use the API properly rather than
backwards as well.

bug #66417